+++ src/main/java/ru/sportlive/mvp/services/UserService.java
@@ -36,7 +36,7 @@ public class UserService {
       return userRepository.findAll();
     }
 
-    public User deposit(Double amount, User user){
+    public User deposit(Integer amount, User user){
         int dep =user.getBalance();
         dep += amount;
         user.setBalance(dep);
